LakeHouse/workspace_dir/requirements.txt
pyspark
minio
psycopg2-binary
delta-spark
confluent_kafka